JWT 토큰 설정

+++ src/main/java/com/example/mic_spring/domain/dto/UserDTO.java
@@ -18,7 +18,7 @@ public class UserDTO {
     private String name;
     private String userId;
     private String email;
-    private Long authority;
+    private Short authority;
     private ZonedDateTime createdAt;
 
     public UserDTO(User user) {
